add gui
