+++ textures/gui.xpm
@@ -1,0 +1,193 @@
+/* XPM */
+static char *_c61625dc50743a787f5642ed2d99834ZvIKwDMhz6tar5Fi[] = {
+/* columns rows colors chars-per-pixel */
+"16 21 166 2 ",
+"   c #1A181C",
+".  c #141114",
+"X  c #2D1A12",
+"o  c #291D1A",
+"O  c #301D14",
+"+  c #341F18",
+"@  c #2B201F",
+"#  c #372117",
+"$  c #3A2115",
+"%  c #35221B",
+"&  c #3C241B",
+"*  c #1E1C23",
+"=  c #232126",
+"-  c #2C2424",
+";  c #232229",
+":  c #312625",
+">  c #362926",
+",  c #342B2C",
+"<  c #3B2E2B",
+"1  c #3F2C25",
+"2  c #3B302E",
+"3  c #32313D",
+"4  c #412517",
+"5  c #442B23",
+"6  c #462E29",
+"7  c #43312C",
+"8  c #52342D",
+"9  c #443632",
+"0  c #4C3630",
+"q  c #473936",
+"w  c #4E3C35",
+"e  c #4A3D3A",
+"r  c #433B3E",
+"t  c #5A3931",
+"y  c #633F37",
+"u  c #633F38",
+"i  c #4D403D",
+"p  c #56443E",
+"a  c #5B453E",
+"s  c #62463F",
+"d  c #6D403E",
+"f  c #75463E",
+"g  c #3F3D47",
+"h  c #5A4747",
+"j  c #5B4B47",
+"k  c #594C49",
+"l  c #5E504F",
+"z  c #5A4E50",
+"x  c #504F59",
+"c  c #5D5052",
+"v  c #484856",
+"b  c #614942",
+"n  c #6C4C45",
+"m  c #744C43",
+"M  c #744E49",
+"N  c #66504B",
+"B  c #6A514A",
+"V  c #765147",
+"C  c #74534C",
+"Z  c #7B554D",
+"A  c #7E594F",
+"S  c #635554",
+"D  c #685650",
+"F  c #645857",
+"G  c #665A5E",
+"H  c #7B5B52",
+"J  c #765D59",
+"K  c #745B53",
+"L  c #6F605C",
+"P  c #70615D",
+"I  c #5B5762",
+"U  c #5C5D6D",
+"Y  c #605D69",
+"T  c #666360",
+"R  c #61616C",
+"E  c #7E6762",
+"W  c #7A6864",
+"Q  c #776B6A",
+"!  c #656576",
+"~  c #825B54",
+"^  c #8B5A56",
+"/  c #896156",
+"(  c #85655C",
+")  c #8C645B",
+"_  c #8E695F",
+"`  c #936C5F",
+"'  c #8A6560",
+"]  c #866964",
+"[  c #8D6B64",
+"{  c #8F6D68",
+"}  c #946D63",
+"|  c #80706D",
+" . c #8E736D",
+".. c #967066",
+"X. c #9A7167",
+"o. c #9A736C",
+"O. c #817373",
+"+. c #86777C",
+"@. c #8A7B78",
+"#. c #947D76",
+"$. c #9A7E78",
+"%. c #947D7A",
+"&. c #A1766D",
+"*. c #A57D74",
+"=. c #8F827E",
+"-. c #99827C",
+";. c #A98276",
+":. c #A9847B",
+">. c #B48A7E",
+",. c #6F7083",
+"<. c #717084",
+"1. c #787481",
+"2. c #7E7883",
+"3. c #757689",
+"4. c #7C7C92",
+"5. c #867C85",
+"6. c #7F8095",
+"7. c #838B8F",
+"8. c #938785",
+"9. c #9F8882",
+"0. c #938E8F",
+"q. c #9E8B8B",
+"w. c #808197",
+"e. c #878A9F",
+"r. c #8A8A9D",
+"t. c #9D9698",
+"y. c #999494",
+"u. c #A08681",
+"i. c #AB8580",
+"p. c #A88D81",
+"a. c #B49188",
+"s. c #BC948C",
+"d. c #B69F9B",
+"f. c #898CA2",
+"g. c #8E91A7",
+"h. c #9395AD",
+"j. c #9498AD",
+"k. c #9899AF",
+"l. c #9194A1",
+"z. c #969AB2",
+"x. c #A7A1A4",
+"c. c #A9A3A4",
+"v. c #AEA9AA",
+"b. c #B3AEB1",
+"n. c #B6B2B5",
+"m. c #BBB6B7",
+"M. c #B9B6B9",
+"N. c #B7B8BA",
+"B. c #BEBABD",
+"V. c #B6A9A5",
+"C. c #C1958F",
+"Z. c #C0998E",
+"A. c #C49C96",
+"S. c #C7A59D",
+"D. c #CAA299",
+"F. c #C8ACA8",
+"G. c #C2A9A5",
+"H. c #D1B3AE",
+"J. c #C2BEC1",
+"K. c #C6C1C4",
+"L. c #CAC5C7",
+"P. c #CBC6C9",
+"I. c #CDC9CC",
+"U. c #D1CDCE",
+"Y. c #D2CED0",
+"T. c #D7D2D6",
+/* pixels */
+"K.I.K.H.M.B.K.K.L.L.U.T.T.U.L.B.",
+"J.J.I.I.I.L.x.O.O.@.8.c.c.x.c.c.",
+"J.J.B.I.I.t.F k q e j P @.c.v.v.",
+"N.b.N.N.c.P i 9 2 2 i > p L m.L.",
+"b.0.0.y.=.N J E N i 9 > o 9 Q V.",
+"N.;   * | V.S.S.a.[ s 7 % @ < i ",
+"g.@ > 2 p.H.D.D.s.>.:.( 7 < > 2 ",
+"6.7 & 1 =.:.Z.D.s.>.;.( p < : 9 ",
+"L & & #.D a } M ) *...C a 7 - l ",
+"k & & d.u. .} 0 2 n _ ..a 9 e G ",
+"6 # + #.F.A.X.] ] H / ` a 7 c G ",
+"& X 4  .s.*.} *.s.:.;.} B n J S ",
+"& X 4  .:.m C o.&.o.} ( ( f b z ",
+"# X $ K V f m / } _ ~ b _ C > F ",
+"k.1.e p &.^ u y ~ m n a Z , . W ",
+"z.z.j.c *.^ ^ Z y 8 8 n ' , ; -.",
+"g.g.j.e.n C M 6 % 5 t A ..h I p.",
+"g.g.g.e.g 2 6 + 4 t m ^ }  .5.u.",
+"f.e.4.6.U p H 8 t m / ) } +.Y r ",
+"6.6.4.! U l } n Z ~ ) { 5.R - * ",
+",.3.<.v 3 I $.{ [ ] O.2.x * * * "
+};
